requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.1.1
psycopg[binary]==3.2.4
gunicorn==22.0.0
python-dotenv==1.0.1
